--- a/.travis.yml
+++ b/.travis.yml
@@ -14,8 +14,9 @@ services:
 
 env:
   matrix:
-  - NODE_VERSION=8
-  - NODE_VERSION=9
+  - NODE_VERSION=8 TEST_SUITE=normal
+  - NODE_VERSION=9 TEST_SUITE=normal
+  - NODE_VERSION="lts/*" TEST_SUITE=performance
   global:
   - COUCH_URL=http://admin:pass@localhost:5984/medic
   - API_URL=http://admin:pass@localhost:5988
@@ -23,6 +24,10 @@ env:
   #AWS Access Key Id and AWS Secret Access Key encrypted by travis to access the S3 buckets where reports and screenshots are saved
   - secure: TR1UN2r3beDtIF+VJpLF2ocTv/uxuOKyVWWhzMLre0ZrrBaIP1sLZV7Z4S/km5M92EfPgGW87BogdWE/R+kTTRiPTCbFB/U/3jFKxEZRXKST66YK5JMwsYqb17UtZtFdqEtO9GGbAVzXwpZfMMoXvlKNrors2W32xBm2uIkOSpI=
   - secure: Q8RH65NClRBryxfvlwHQjeR4wGs+GXeUBRIBd1kspAM7Uv+5K3iP+q/TPCIrTL/OpJ5tyCYfeq5hKIFFwrY3JikNXnyHICOjXPE6zSLUm8E8NHaP/orPNWjze2x4yDSRCifqzr2ZiXVq9sxlZfNbZ9eyJxpPFTYDSpuN9T5UEE0=
+
+matrix:
+  allow_failures:
+    - env: NODE_VERSION="lts/*" TEST_SUITE=performance
 
 before_install:
   # start couchdb 2.x docker instance
@@ -54,7 +59,9 @@ before_script:
   - until nc -z localhost 4444; do sleep 1; done
 
 script:
-  - node --stack_size=10000 `which grunt` ci
+  - |
+    [[ "$TEST_SUITE" != 'normal' ]]      || node --stack_size=10000 `which grunt` ci
+    [[ "$TEST_SUITE" != 'performance' ]] || node --stack_size=10000 `which grunt` ci-performance
 
 after_script:
 - |