--- a/tests/protractor/e2e.tests-and-services.conf.js
+++ b/tests/protractor/e2e.tests-and-services.conf.js
@@ -1,0 +1,2 @@
+const BaseConfig = require('./base.conf.js');
+module.exports = new BaseConfig('e2e', { manageServices:true });