--- a/tests/protractor/e2e.tests-only.conf.js
+++ b/tests/protractor/e2e.tests-only.conf.js
@@ -1,3 +1,2 @@
 const BaseConfig = require('./base.conf.js');
-
-module.exports = new BaseConfig();
+module.exports = new BaseConfig('e2e');