--- a/tests/protractor/e2e/api/controllers/sms-gateway.spec.js
+++ b/tests/protractor/e2e/api/controllers/sms-gateway.spec.js
@@ -1,7 +1,9 @@
-const utils = require('../../../utils');
-const uuid = require('uuid/v4');
+const testUtils = require('../../../utils');
+const gatewayApiUtils = require('../../../gateway-api.utils');
+const api = gatewayApiUtils.api;
+const db = gatewayApiUtils.db;
+const setup = gatewayApiUtils.setup;
 
-const CHW_CONTACT_NUMBER = '+32049832049';
 
 describe('/sms', function() {
 
@@ -9,7 +11,7 @@ describe('/sms', function() {
 
     it('should respond with valid JSON', function() {
       // when
-      return get()
+      return api.get()
         .then(response => {
           // then
           expect(response['medic-gateway']).toBe(true);
@@ -20,27 +22,19 @@ describe('/sms', function() {
 
   describe('POST', function() {
 
-    afterAll(() => utils.revertDb());
+    afterAll(() => testUtils.revertDb());
 
-    afterEach(() => {
-      // delete WO and WT messages
-      return allMessageDocs()
-        .then(docs => docs.map(doc => {
-          doc._deleted = true;
-          return doc;
-        }))
-        .then(docs => utils.db.bulkDocs(docs));
-    });
+    afterEach(gatewayApiUtils.cleanUp);
 
     describe('for webapp-terminating message processing', function() {
 
       it('should accept requests with missing fields', function() {
-        return post({})
-          .then(expectResponse({ messages:[] }));
+        return api.post({})
+          .then(assert.response({ messages:[] }));
       });
 
       it('should save supplied messages to DB', function() {
-        return postMessages(
+        return api.postMessages(
           {
             id: 'test-sms-1',
             from: '+111-test',
@@ -56,9 +50,9 @@ describe('/sms', function() {
             sms_received: 1511189020577,
           }
         )
-          .then(expectResponse({ messages:[] }))
+          .then(assert.response({ messages:[] }))
 
-          .then(() => expectMessagesInDb('test 1', 'test 2'));
+          .then(() => assert.messagesInDb('test 1', 'test 2'));
       });
 
     });
@@ -66,237 +60,140 @@ describe('/sms', function() {
     describe('for webapp-originating message processing', function() {
 
       it('should update message in DB when status update is received', function() {
-        return saveWoMessage('abc-123', 'hello from webapp')
-          .then(() => expectMessageWithoutState('abc-123'))
+        return setup.saveWoMessage('abc-123', 'hello from webapp')
+          .then(() => assert.messageWithoutState('abc-123'))
 
-          .then(() => postStatus('abc-123', 'SENT'))
-          .then(expectResponse({ messages:[] }))
+          .then(() => api.postStatus('abc-123', 'SENT'))
+          .then(assert.response({ messages:[] }))
 
-          .then(() => expectMessageState('abc-123', 'sent'));
+          .then(() => assert.messageState('abc-123', 'sent'));
       });
 
       it('should update message multiple times when multiple status updates for the same message are received', function() {
-        return saveWoMessage('abc-123', 'hello again')
-          .then(() => expectMessageWithoutState('abc-123'))
+        return setup.saveWoMessage('abc-123', 'hello again')
+          .then(() => assert.messageWithoutState('abc-123'))
 
-          .then(() => postStatuses(
+          .then(() => api.postStatuses(
               { id:'abc-123', status:'SENT' },
               { id:'abc-123', status:'DELIVERED' }))
-          .then(expectResponse({ messages:[] }))
+          .then(assert.response({ messages:[] }))
 
-          .then(() => expectMessageStates({ id:'abc-123', states:['sent', 'delivered'] }));
+          .then(() => assert.messageStates({ id:'abc-123', states:['sent', 'delivered'] }));
       });
 
       it('should not save a status update again if it\'s been seen before', function() {
-        return saveWoMessage('abc-123', 'hello again')
-          .then(() => expectMessageWithoutState('abc-123'))
+        return setup.saveWoMessage('abc-123', 'hello again')
+          .then(() => assert.messageWithoutState('abc-123'))
 
-          .then(() => postStatus('abc-123', 'SENT'))
-          .then(expectResponse({ messages:[] }))
-          .then(() => expectMessageStates({ id:'abc-123', states:['sent'] }))
+          .then(() => api.postStatus('abc-123', 'SENT'))
+          .then(assert.response({ messages:[] }))
+          .then(() => assert.messageStates({ id:'abc-123', states:['sent'] }))
 
-          .then(() => postStatus('abc-123', 'SENT'))
-          .then(expectResponse({ messages:[] }))
-          .then(() => expectMessageStates({ id:'abc-123', states:['sent'] }));
+          .then(() => api.postStatus('abc-123', 'SENT'))
+          .then(assert.response({ messages:[] }))
+          .then(() => assert.messageStates({ id:'abc-123', states:['sent'] }));
       });
 
     });
 
     it('should still save messages when a status update for an unknown message is received', function() {
-        return saveWoMessage('abc-123', 'hello again')
-          .then(() => expectMessageWithoutState('abc-123'))
+        return setup.saveWoMessage('abc-123', 'hello again')
+          .then(() => assert.messageWithoutState('abc-123'))
 
-          .then(() => postStatuses(
+          .then(() => api.postStatuses(
               { id:'abc-123', status:'SENT' },
               { id:'def-456', status:'DELIVERED' }))
-          .then(expectResponse({ messages:[] }))
-          .then(() => expectMessageStates({ id:'abc-123', states:['sent'] }));
+          .then(assert.response({ messages:[] }))
+          .then(() => assert.messageStates({ id:'abc-123', states:['sent'] }));
     });
 
     it('should still save messages when an unrecognised status update is received', function() {
-        return saveWoMessage('abc-123', 'hello again')
-          .then(() => expectMessageWithoutState('abc-123'))
+        return setup.saveWoMessage('abc-123', 'hello again')
+          .then(() => assert.messageWithoutState('abc-123'))
 
-          .then(() => postStatus('abc-123', 'WTF'))
-          .then(expectResponse({ messages:[] }))
-          .then(() => expectMessageStates({ id:'abc-123', states:['unrecognised'] }));
+          .then(() => api.postStatus('abc-123', 'WTF'))
+          .then(assert.response({ messages:[] }))
+          .then(() => assert.messageStates({ id:'abc-123', states:['unrecognised'] }));
     });
 
   });
 });
 
-function get() {
-  return utils.request({
-    path: '/api/sms',
-    method: 'GET',
-    headers: { 'Content-Type':'application/json' },
-  });
-}
+const assert = {
 
-function post(body) {
-  return utils.request({
-    path: '/api/sms',
-    method: 'POST',
-    headers: { 'Content-Type':'application/json' },
-    body: body,
-  });
-}
+  response: (expected) => {
+    return response => expect(response).toEqual(expected);
+  },
 
-function postMessages(...messages) {
-  return post({ messages });
-}
+  messagesInDb: (...expectedContents) => {
+    expectedContents = JSON.stringify(expectedContents);
 
-function expectResponse(expected) {
-  return response => expect(response).toEqual(expected);
-}
+    return new Promise((resolve, reject) => {
+      const endTime = Date.now() + 10000;
 
-function expectMessagesInDb(...expectedContents) {
-  expectedContents = JSON.stringify(expectedContents);
+      check();
 
-  return new Promise((resolve, reject) => {
-    const endTime = Date.now() + 10000;
+      function check() {
+        db.getMessageContents()
+          .then(actualContents => {
+            actualContents = JSON.stringify(actualContents);
 
-    check();
-
-    function check() {
-      getMessageContents()
-        .then(actualContents => {
-          actualContents = JSON.stringify(actualContents);
-
-          if(actualContents === expectedContents) {
-            resolve();
-          } else if(Date.now() < endTime) {
-            setTimeout(check, 100);
-          } else {
-            reject(`Expected:\n      ${actualContents}\n    to equal:\n      ${expectedContents}`);
-          }
-        })
-        .catch(reject);
-    }
-  });
-}
-
-function saveWoMessage(id, content) {
-  return saveWoMessages({ id, content });
-}
-
-function saveWoMessages(...details) {
-  return utils.db.bulkDocs(details.map(d => createWoMessage(d.id, d.content)));
-}
-
-function createWoMessage(id, content) {
-  const task = {
-    messages: [
-      {
-        from: '+123',
-        sent_by: 'some name',
-        to: CHW_CONTACT_NUMBER,
-        contact: {},
-        message: content,
-        uuid: id,
+            if(actualContents === expectedContents) {
+              resolve();
+            } else if(Date.now() < endTime) {
+              setTimeout(check, 100);
+            } else {
+              reject(`Expected:\n      ${actualContents}\n    to equal:\n      ${expectedContents}`);
+            }
+          })
+          .catch(reject);
       }
-    ]
-  };
+    });
+  },
 
-  const messageDoc = {
-    _id: uuid(),
-    errors: [],
-    form: null,
-    from: '+123',
-    reported_date: 1520416423761,
-    tasks: [ task ],
-    kujua_message: true,
-    type: 'data_record',
-    sent_by: 'some name',
-  };
+  messageWithoutState: (id) => {
+    return assert.messagesWithoutState(id);
+  },
 
-  return messageDoc;
-}
+  messagesWithoutState: (...ids) => {
+    return assert.messageStates(...ids.map(id => ({ id })));
+  },
 
-function postStatus(messageId, newStatus) {
-  return postStatuses({ id:messageId, status:newStatus });
-}
+  messageState: (id, state) => {
+    return assert.messageStates({ id, state });
+  },
 
-function postStatuses(...updates) {
-  return post({ updates });
-}
-
-function expectMessageWithoutState(id) {
-  return expectMessagesWithoutState(id);
-}
-
-function expectMessagesWithoutState(...ids) {
-  return expectMessageStates(...ids.map(id => ({ id })));
-}
-
-function expectMessageState(id, state) {
-  return expectMessageStates({ id, state });
-}
-
-function expectMessageStates(...expectedStates) {
-  expectedStates.forEach(expectation => {
-    if(expectation.state) {
-      expectation.states = [ expectation.state ];
-      delete expectation.state;
-    }
-  });
-
-  expectedStates = JSON.stringify(expectedStates);
-
-  return new Promise((resolve, reject) => {
-    const endTime = Date.now() + 10000;
-
-    check();
-
-    function check() {
-      getMessageStates()
-        .then(actualStates => {
-          actualStates = JSON.stringify(actualStates);
-
-          if(actualStates === expectedStates) {
-            resolve();
-          } else if(Date.now() < endTime) {
-            setTimeout(check, 100);
-          } else {
-            reject(`Expected:\n      ${actualStates}\n    to equal:\n      ${expectedStates}`);
-          }
-        })
-        .catch(reject);
-    }
-  });
-}
-
-function getMessageStates() {
-  return allMessageDocs()
-    .then(docs => docs.reduce((acc, doc) => {
-      doc.tasks.forEach(task => task.messages.forEach(m => {
-        const states = task.state_history &&
-                       task.state_history.map(h => h.state);
-        acc.push({ id:m.uuid, states });
-      }));
-      return acc;
-    }, []));
-}
-
-function getMessageContents() {
-  return allMessageDocs()
-    .then(docs => docs.reduce((acc, doc) => {
-
-      if(doc.kujua_message) {
-        doc.tasks.forEach(task =>
-            task.messages.forEach(m => acc.push(m.message)));
+  messageStates: (...expectedStates) => {
+    expectedStates.forEach(expectation => {
+      if(expectation.state) {
+        expectation.states = [ expectation.state ];
+        delete expectation.state;
       }
+    });
 
-      if(doc.sms_message) {
-        acc.push(doc.sms_message.message);
+    expectedStates = JSON.stringify(expectedStates);
+
+    return new Promise((resolve, reject) => {
+      const endTime = Date.now() + 10000;
+
+      check();
+
+      function check() {
+        db.getMessageStates()
+          .then(actualStates => {
+            actualStates = JSON.stringify(actualStates);
+
+            if(actualStates === expectedStates) {
+              resolve();
+            } else if(Date.now() < endTime) {
+              setTimeout(check, 100);
+            } else {
+              reject(`Expected:\n      ${actualStates}\n    to equal:\n      ${expectedStates}`);
+            }
+          })
+          .catch(reject);
       }
+    });
+  },
 
-      return acc;
-    }, []));
-}
-
-function allMessageDocs() {
-  return utils.db.query('medic-client/messages_by_contact_date',
-      { reduce:false, include_docs:true })
-    .then(res => res.rows.map(row => row.doc));
-}
+};