--- a/tests/protractor/performance.tests-and-services.conf.js
+++ b/tests/protractor/performance.tests-and-services.conf.js
@@ -1,0 +1,3 @@
+const BaseConfig = require('./base.conf.js');
+
+module.exports = new BaseConfig('performance', { manageServices:true });